Rename repository and sub directory to "github-v4"
Following convention
