+++ main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/hashicorp/terraform/plugin"
-	"github.com/patrickmarabeas/terraform-provider-githubv4/github"
+	"github.com/patrickmarabeas/terraform-provider-github-v4/github-v4"
 )
 
 func main() {
